follow flask docs on circular imports
http://flask.pocoo.org/docs/1.0/patterns/packages/
